Added "Game Over" message to the gui. Waiting until key was pressed to leave game was creating issues so used timing to end game after a few seconds. Increased font of score text. Changed background color to light blue.
Was getting the following error: Warning:(74, 9) 'while' statement cannot complete without throwing an exception

The warning I was seeing, ""while statement cannot complete without throwing an exception", meant that the while loop in the run() method did not have any way to terminate except by throwing an exception. This can occur if the while loop condition is always true, and there is no code inside the loop that can change the condition to false.

 To fix this warning, I modified the while loop to include a boolean variable.